+++ .gitignore
@@ -51,4 +51,4 @@ test-output
 
 # LangGraph API
 .langgraph_api
-apps/mastra.db
+apps/mastra.db*
