+++ apps/agentic-server/src/mastra/tools/executeSwap.ts
@@ -1,0 +1,9 @@
+import { createTool } from '@mastra/core/tools';
+import { z } from 'zod';
+
+export const executeSwap = createTool({
+  id: 'executeSwap',
+  inputSchema: z.object({}),
+  description:
+    'Sends a transaction which executes the swap the user has confirmed.',
+});
