+++ apps/agentic-server/src/mastra/tools/getAddress.ts
@@ -1,0 +1,8 @@
+import { createTool } from '@mastra/core/tools';
+import { z } from 'zod';
+
+export const getAddress = createTool({
+  id: 'getAddress',
+  inputSchema: z.object({}),
+  description: 'Returns the user address across all EVM chains',
+});
